Guessing Game
Those who like Guessing games maybe try this! This has 26 Levels, if you want to update it permanently, give an update on the source code and Message me through Github Or My Gmail That is in the Source Code! I will be posting How To Use All Statements and functions that are in the source code! Want to be a developer without a salary? Well Then Come On In And Join Me! Text Me Through GitHub Or Gmail So I Can give you a form you have to fill. Im Giving The Orginal Version Of The File , ill give it through Mediafire, the format of github is quite wrong so I'm giving the source code out, the original one.